jgmenu integration
Added jgmenu integration. Previous configuration was not correct, was using jgmenu and jgmenu_run. Now it has a proper menu, based on the archlabs theme (inclued via jgmenu init -i). Tried to make theme the same as gtk but is not working, will revisit in the future. small correction to i3lock.