sounds!
i3 now have a sound ""system"". its now possible to play sounds when something is completed or the user is asked for something ie pkexec passw. this is managed within i3 config. the i3sound is just a little script to play them. currently the sounds are from ubuntu yaru theme. the script is using paplay from libpulse.